+++ src/main/scala/io/rml/framework/flink/source/JSONStream.scala
@@ -1,7 +1,7 @@
 package io.rml.framework.flink.source
 
 import io.rml.framework.core.internal.Logging
-import io.rml.framework.core.model.{FileStream, KafkaStream, Literal, StreamDataSource, TCPSocketStream}
+import io.rml.framework.core.model.{FileStream, KafkaStream, StreamDataSource, TCPSocketStream}
 import io.rml.framework.core.vocabulary.RMLVoc
 import io.rml.framework.flink.item.Item
 import io.rml.framework.flink.item.json.JSONItem
@@ -9,27 +9,23 @@ import org.apache.flink.api.common.serialization.SimpleStringSchema
 import org.apache.flink.api.scala._
 import org.apache.flink.streaming.api.scala.{DataStream, StreamExecutionEnvironment}
 
-case class JSONStream(stream: DataStream[Iterable[Item]]) extends Stream
+case class JSONStream(val stream: DataStream[Iterable[Item]]) extends Stream
 
 object JSONStream extends Logging {
-  val DEFAULT_PATH_OPTION: String = Source.DEFAULT_ITERATOR_MAP(RMLVoc.Class.JSONPATH).getOrElse("$")
+  val DEFAULT_PATH_OPTION: String = Source.DEFAULT_ITERATOR_MAP(RMLVoc.Class.JSONPATH)
 
-  def apply(source: StreamDataSource, jsonPaths: List[Option[Literal]])(implicit env: StreamExecutionEnvironment): Stream = {
-    val jsonPathStrings = jsonPaths.map({
-      case Some(x) => x.toString
-      case _ => DEFAULT_PATH_OPTION
-    })
-      .distinct
+  def apply(source: StreamDataSource, jsonPaths: List[String])(implicit env: StreamExecutionEnvironment): Stream = {
 
     source match {
-      case tcpStream: TCPSocketStream => fromTCPSocketStream(tcpStream, jsonPathStrings)
-      case fileStream: FileStream => fromFileStream(fileStream.path, jsonPathStrings)
-      case kafkaStream: KafkaStream => fromKafkaStream(kafkaStream, jsonPathStrings)
+      case tcpStream: TCPSocketStream => fromTCPSocketStream(tcpStream, jsonPaths)
+      case fileStream: FileStream => fromFileStream(fileStream.path, jsonPaths)
+      case kafkaStream: KafkaStream => fromKafkaStream(kafkaStream, jsonPaths)
     }
   }
 
   def fromTCPSocketStream(tCPSocketStream: TCPSocketStream, jsonPaths: List[String])(implicit env: StreamExecutionEnvironment): JSONStream = {
-    val stream: DataStream[Iterable[Item]] = StreamUtil.createTcpSocketSource(tCPSocketStream)
+    val socketStream = StreamUtil.paralleliseOverSlots(StreamUtil.createTcpSocketSource(tCPSocketStream))
+    val stream: DataStream[Iterable[Item]] = socketStream
       .map { item =>
         JSONItem.fromStringOptionableList(item, jsonPaths)
       }
@@ -49,7 +45,8 @@ object JSONStream extends Logging {
     val consumer = kafkaStream.getConnectorFactory.getSource(kafkaStream.topic, new SimpleStringSchema(), properties)
 
     logDebug(consumer.getProducedType.toString)
-    val stream: DataStream[Iterable[Item]] = env.addSource(consumer)
+    val parallelStream = StreamUtil.paralleliseOverSlots(env.addSource(consumer))
+    val stream: DataStream[Iterable[Item]] = parallelStream
       .map { item =>
         JSONItem.fromStringOptionableList(item, jsonPaths)
       }